NaturalHistory_Transformer_API_v1.0/requirements.txt
Flask==2.2.2
Flask==2.2.2
Werkzeug==2.2.2
torch==2.2.2
transformers==4.38.2
requests==2.31.0
python-dotenv==0.21.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==20.1.0
numpy==1.24.3
pandas==2.0.3
